ADR-002 stage 2.6 (PR1/3): split core chat/node intents out of register_canvas
register_canvas (backend/canvas.py) was one ~1572-line function wiring
~89 WS intents/topics onto a SessionBus. This slice relocates the
highest-value, highest-risk chunk - the 14 intents + 2 helper closures
making up the Composer's real Send/Regenerate actions and generic node
CRUD - into 4 new modules under backend/api/, each exporting one
register_*_intents(bus, document, ...) function:

- intents_nodes.py: generic node/edge CRUD (addNode..removeEdges)
- intents_conversation.py: ConversationNode lifecycle
- intents_chat.py: sendMessage/regenerateResponse/cancelChatRequest -
  the largest and most load-bearing closures in the file
- intents_chat_image.py: image generation/regeneration, kept separate
  from intents_chat.py since it runs on AgentDispatcher's independent
  "image" run kind, a genuine different feature not just a size trim

backend/api/_shared.py factors out publish_scene/publish_token_counter,
the two closures the old register_canvas shared across ~60 of its ~89
intents - each new module now gets its own equivalent via a factory
function instead.

Pure code motion - every relocated closure body verified byte-for-byte
identical against `git show origin/main:backend/canvas.py` before
shipping, not just eyeballed. The one deliberate wrinkle:
intents_chat.py imports _history_token_text back from backend.canvas at
call time (a function-body-local import, not module-level) to resolve
a genuine circular import - backend.canvas now imports register_chat_
intents FROM this module, while this module needs a helper still
defined in backend.canvas; resolved by import ORDER (the new imports
sit after canvas.py's own helper defs, not with its other top-of-file
imports) rather than avoided.

register_canvas itself is now a thin orchestrator for this slice's
scope: constructs the shared SceneDocument, registers the topics with
no natural feature home, calls the 4 new functions. The ~75 remaining
intents (web research, artifact, chart, gitlink, pycoder, code sandbox,
groups, view, grid) are untouched, deferred to PR2/PR3.

Two real mistakes found and fixed during self-review, both belonging to
the same "zero test signal" bug class since neither is caught by
CI (no lint step exists) or by the test suite (nothing calls dead
code, so nothing fails when it's left behind):
1. The OLD bus.register_intent(...) call sites for these 14 intents
   lived in a separate deferred "tail block" far below their own
   closures - initially left behind referencing now-deleted names (a
   real NameError, caught immediately by the full test suite failing
   everywhere - not silent).
2. moveNode/moveNodes/removeNodes/connectNodes/removeEdges were copied
   into the new intents_nodes.py, but the ORIGINAL closures in
   canvas.py were never deleted - orphaned, unreachable dead code with
   zero test signal (nothing calls an unregistered closure), found via
   manual grep sweep, not by any test failure.

A follow-up adversarial review (2 independent reviewers + synthesis)
found two more instances of the same underlying risk class before this
shipped: 5 orphaned module-level imports left in canvas.py (base64,
parse_response + 3 PLACEHOLDER_* constants - all runtime-harmless,
unused bindings, but genuine leftover artifacts) and 5 comments whose
"above"/"this scope" cross-references went stale once their target
closures moved to a different file. Both fixed; a repo-wide pyflakes
sweep after the fixes confirms nothing else was missed (the remaining
pyflakes hits in canvas.py are its own pre-existing, deliberate domain-
model re-exports from ADR-002 stage 2.2, untouched by this PR).

Full suite: 1224 passed (unchanged count - pure behavior preservation).
